Move back event-emitter since it’s used in router
