+++ packages/next-server/lib/event-emitter.js
