+++ packages/next-server/lib/router/router.js
@@ -1,7 +1,7 @@
 /* global __NEXT_DATA__ */
 
 import { parse, format } from 'url'
-import EventEmitter from '../EventEmitter'
+import EventEmitter from '../event-emitter'
 import shallowEquals from './shallow-equals'
 import { loadGetInitialProps, getURL } from '../utils'
