+++ packages/next/client/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom'
 import HeadManager from './head-manager'
 import { createRouter } from 'next/router'
-import EventEmitter from './event-emitter'
+import EventEmitter from 'next-server/dist/lib/event-emitter'
 import {loadGetInitialProps, getURL} from 'next-server/dist/lib/utils'
 import PageLoader from './page-loader'
 import * as asset from 'next-server/asset'
